requirements.txt
Django>=4.2.0
channels==4.0.0
channels-redis==4.1.0
django-crispy-forms>=2.0
Pillow>=10.0.0
python-decouple>=3.8
django-allauth>=0.54.0
django-notifications-hq>=1.8.0
redis>=5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
